# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Rezina Civic reporting app
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReportCategory = typeof ReportCategory[keyof typeof ReportCategory];


export const ReportCategory = {
  parking: 'parking',
  accident_traffic: 'accident_traffic',
  sewage: 'sewage',
  garbage_road: 'garbage_road',
  tree: 'tree',
  electricity: 'electricity',
  road: 'road',
  water: 'water',
  garbage: 'garbage',
  lighting: 'lighting',
  vandalism: 'vandalism',
  stray_animals: 'stray_animals',
  other: 'other',
} as const;

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  pending: 'pending',
  resolved: 'resolved',
} as const;

export interface Report {
  id: number;
  title: string;
  description: string;
  category: ReportCategory;
  latitude: number;
  longitude: number;
  /** @nullable */
  photoBase64?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  reporterName?: string | null;
  /** @nullable */
  reporterEmail?: string | null;
  status: ReportStatus;
  resolvedVotes: number;
  createdAt: string;
}

export type ReportInputCategory = typeof ReportInputCategory[keyof typeof ReportInputCategory];


export const ReportInputCategory = {
  parking: 'parking',
  accident_traffic: 'accident_traffic',
  sewage: 'sewage',
  garbage_road: 'garbage_road',
  tree: 'tree',
  electricity: 'electricity',
  road: 'road',
  water: 'water',
  garbage: 'garbage',
  lighting: 'lighting',
  vandalism: 'vandalism',
  stray_animals: 'stray_animals',
  other: 'other',
} as const;

export interface ReportInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  description: string;
  category: ReportInputCategory;
  latitude: number;
  longitude: number;
  /** @nullable */
  photoBase64?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  reporterName?: string | null;
  /** @nullable */
  reporterEmail?: string | null;
}

export interface VoteInput {
  /** @nullable */
  voterFingerprint?: string | null;
}

export interface ErrorResponse {
  error: string;
}