# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Rezina Civic reporting app
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reports
    description: Urban problem reports
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all reports
      description: Returns all urban problem reports with their map pins
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"

    post:
      operationId: createReport
      tags: [reports]
      summary: Create a new report
      description: Submit a new urban problem report with optional photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a single report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reports/{id}/vote-resolved:
    post:
      operationId: voteResolved
      tags: [reports]
      summary: Vote that a report has been resolved
      description: When 3 or more users vote resolved, the report is marked as resolved
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteInput"
      responses:
        "200":
          description: Vote registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Already voted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Report:
      type: object
      required:
        - id
        - title
        - description
        - category
        - latitude
        - longitude
        - status
        - resolvedVotes
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum:
            - parking
            - accident_traffic
            - sewage
            - garbage_road
            - tree
            - electricity
            - road
            - water
            - garbage
            - lighting
            - vandalism
            - stray_animals
            - other
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        photoBase64:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        reporterName:
          type: ["string", "null"]
        reporterEmail:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, resolved]
        resolvedVotes:
          type: integer
        createdAt:
          type: string

    ReportInput:
      type: object
      required:
        - title
        - description
        - category
        - latitude
        - longitude
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
          minLength: 10
        category:
          type: string
          enum:
            - parking
            - accident_traffic
            - sewage
            - garbage_road
            - tree
            - electricity
            - road
            - water
            - garbage
            - lighting
            - vandalism
            - stray_animals
            - other
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        photoBase64:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        reporterName:
          type: ["string", "null"]
        reporterEmail:
          type: ["string", "null"]

    VoteInput:
      type: object
      properties:
        voterFingerprint:
          type: ["string", "null"]

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string